Move association+rbac related stuff in a dedicated file
