+++ pkg/controller/common/association/association.go
@@ -8,10 +8,7 @@ import (
 	commonv1 "github.com/elastic/cloud-on-k8s/pkg/apis/common/v1"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/events"
 	"github.com/elastic/cloud-on-k8s/pkg/utils/k8s"
-	"github.com/elastic/cloud-on-k8s/pkg/utils/rbac"
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
 )
@@ -49,38 +46,4 @@ func IsConfiguredIfSet(associated commonv1.Associated, r record.EventRecorder) b
 		return false
 	}
 	return true
-}
-
-type Unbinder interface {
-	Unbind(associated commonv1.Associated) error
-}
-
-// CheckAndUnbind checks if a reference is allowed and unbind the association if it is not the case
-func CheckAndUnbind(
-	accessReviewer rbac.AccessReviewer,
-	associated commonv1.Associated,
-	object runtime.Object,
-	unbinder Unbinder,
-) (bool, error) {
-	allowed, err := accessReviewer.AccessAllowed(associated.ServiceAccountName(), associated.GetNamespace(), object)
-	if err != nil {
-		return false, err
-	}
-	if !allowed {
-		metaObject, err := meta.Accessor(object)
-		if err != nil {
-			return false, nil
-		}
-		log.Info("Association not allowed",
-			"associated_kind", associated.GetObjectKind().GroupVersionKind().Kind,
-			"associated_name", associated.GetName(),
-			"associated_namespace", associated.GetNamespace(),
-			"serviceAccount", associated.ServiceAccountName(),
-			"remote_type", object.GetObjectKind().GroupVersionKind().Kind,
-			"remote_name", metaObject.GetNamespace(),
-			"remote_namespace", metaObject.GetName(),
-		)
-		return false, unbinder.Unbind(associated)
-	}
-	return true, nil
 }
