+++ pkg/controller/common/association/association_test.go
@@ -9,12 +9,9 @@ import (
 
 	apmv1 "github.com/elastic/cloud-on-k8s/pkg/apis/apm/v1"
 	commonv1 "github.com/elastic/cloud-on-k8s/pkg/apis/common/v1"
-	esv1 "github.com/elastic/cloud-on-k8s/pkg/apis/elasticsearch/v1"
 	"github.com/elastic/cloud-on-k8s/pkg/utils/k8s"
-	"github.com/elastic/cloud-on-k8s/pkg/utils/rbac"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 )
 
 func TestGetCredentials(t *testing.T) {
@@ -102,94 +99,6 @@ func TestGetCredentials(t *testing.T) {
 			}
 			if gotPassword != tt.wantPassword {
 				t.Errorf("getCredentials() gotPassword = %v, want %v", gotPassword, tt.wantPassword)
-			}
-		})
-	}
-}
-
-type fakeAccessReviewer struct {
-	allowed bool
-	err     error
-}
-
-func (f *fakeAccessReviewer) AccessAllowed(_ string, _ string, _ runtime.Object) (bool, error) {
-	return f.allowed, f.err
-}
-
-type fakeUnbinder struct {
-	called bool
-}
-
-func (f *fakeUnbinder) Unbind(associated commonv1.Associated) error {
-	f.called = true
-	return nil
-}
-
-func TestCheckAndUnbind(t *testing.T) {
-	apmServer := &apmv1.ApmServer{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "apm-server-sample",
-			Namespace: "apmserver-ns",
-		},
-		Spec: apmv1.ApmServerSpec{},
-	}
-	es := &esv1.Elasticsearch{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "es-sample",
-			Namespace: "es-ns",
-		},
-	}
-
-	type args struct {
-		accessReviewer rbac.AccessReviewer
-		associated     commonv1.Associated
-		object         runtime.Object
-		unbinder       fakeUnbinder
-	}
-	tests := []struct {
-		name                                  string
-		args                                  args
-		want, wantErr, wantFakeUnbinderCalled bool
-	}{
-		{
-			name: "Association not allowed, ensure unbinder is called",
-			args: args{
-				associated: apmServer,
-				object:     es,
-				accessReviewer: &fakeAccessReviewer{
-					allowed: false,
-				},
-				unbinder: fakeUnbinder{},
-			},
-			wantFakeUnbinderCalled: true,
-			want:                   false,
-		},
-		{
-			name: "Association allowed, ensure unbinder is not called",
-			args: args{
-				associated: apmServer,
-				object:     es,
-				accessReviewer: &fakeAccessReviewer{
-					allowed: true,
-				},
-				unbinder: fakeUnbinder{},
-			},
-			wantFakeUnbinderCalled: false,
-			want:                   true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := CheckAndUnbind(tt.args.accessReviewer, tt.args.associated, tt.args.object, &tt.args.unbinder)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("CheckAndUnbind() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("CheckAndUnbind() = %v, want %v", got, tt.want)
-			}
-			if tt.args.unbinder.called != tt.wantFakeUnbinderCalled {
-				t.Errorf("fakeUnbinder.called = %v, want %v", tt.args.unbinder.called, tt.wantFakeUnbinderCalled)
 			}
 		})
 	}
